# api/openapi.yaml
openapi: 3.0.3
info:
  title: Go Backend Template API
  description: REST API for go-backend-template
  version: "1.0"
servers:
  - url: /api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, approved_id]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        approved_id: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, token_type]
      properties:
        token: { type: string }
        token_type: { type: string }
    UserResponse:
      type: object
      required: [id, email, first_name, is_active, roles, created_at]
      properties:
        id: { type: string }
        email: { type: string }
        first_name: { type: string }
        is_active: { type: boolean }
        roles:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
    CreateTodoRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 255 }
        description: { type: string, nullable: true }
        due_date: { type: string, format: date-time, nullable: true }
    UpdateTodoRequest:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 255, nullable: true }
        description: { type: string, nullable: true }
        is_completed: { type: boolean, nullable: true }
        due_date: { type: string, format: date-time, nullable: true }
    TodoResponse:
      type: object
      required: [id, user_id, title, is_completed, created_at, updated_at]
      properties:
        id: { type: string }
        user_id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        is_completed: { type: boolean }
        due_date: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ApprovedUserRequest:
      type: object
      required: [email, first_name]
      properties:
        email: { type: string, format: email }
        first_name: { type: string, minLength: 1 }
    BulkApprovedUserRequest:
      type: object
      required: [users]
      properties:
        users:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ApprovedUserRequest' }
    ApprovedUserResponse:
      type: object
      required: [id, email, first_name, created_at, updated_at]
      properties:
        id: { type: string }
        email: { type: string }
        first_name: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
  /auth/register:
    post:
      operationId: RegisterUser
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
  /auth/login:
    post:
      operationId: LoginUser
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
  /me:
    get:
      operationId: GetCurrentUser
      tags: [auth]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /todos:
    get:
      operationId: ListTodos
      tags: [todos]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TodoResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
    post:
      operationId: CreateTodo
      tags: [todos]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTodoRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TodoResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
  /todos/{id}:
    get:
      operationId: GetTodo
      tags: [todos]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TodoResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
    patch:
      operationId: UpdateTodo
      tags: [todos]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTodoRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TodoResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
    delete:
      operationId: DeleteTodo
      tags: [todos]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
  /admin/approved-users:
    get:
      operationId: ListApprovedUsers
      tags: [admin]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ApprovedUserResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
    post:
      operationId: CreateApprovedUser
      tags: [admin]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovedUserRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApprovedUserResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
  /admin/approved-users/bulk:
    post:
      operationId: BulkCreateApprovedUsers
      tags: [admin]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkApprovedUserRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ApprovedUserResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
  /admin/approved-users/{id}:
    delete:
      operationId: DeleteApprovedUser
      tags: [admin]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }